feat: change token based password recovery to code based recovery
BREAKING CHANGES: password recovery now use passcode instead of SHA
token. The previous SHA token based password recovery required the
user to click the link sent to his email, something which was difficult
if the user use email in a separate device so now by using passcode
the implementor can allow user to copy the 6 digit code sent to his email
without relying on the device which receive the email.